requirements.txt
Django==1.3.1
South==0.7.3
distribute==0.6.15
django-debug-toolbar==0.9.4
factory-boy==1.1.3
ipython==0.12
wsgiref==0.1.2
yolk==0.4.3
